==> HMSC/3_VariancePartitioningFigure.R <==
setwd()

library(Hmsc)

#load Hmsc model generated by script 2.
load("models_thin_1000_samples_250_chains_4.Rdata")

library(Hmsc)

VPsummary = matrix(NA, nrow = 0, ncol = 8)

modeltype = c("OTU models", "Haplotype models")
plot_titles = c("mOTU occurrence", "mOTU abundance", "Haplotype occurrence", "Haplotype abundance")

#Variance partitioning summary figures with all fixed and random effects illustrated separately.
pdf("VarPart_SuppFig8.pdf", width = 8, height = 8)
par(mfrow = c(2, 2), mar = c(1, 4, 3, 1), lwd = 0.1)

for(j in 1:4)
  {
    m = models[[j]]
    preds = computePredictedValues(m)
    VP = computeVariancePartitioning(m)
    MF = evaluateModelFit(hM=m, predY=preds)
    if(j == 1 | j == 3)
    {R2 = MF$TjurR2}
    if(j == 2 | j == 4)
    {R2 = MF$R2}
    for(k in 1:m$ns){
      VP$vals[,k] = R2[k]*VP$vals[,k]
    }
    VP$vals = VP$vals[,order(R2, decreasing = TRUE)]
    VPsummary = rbind(VPsummary, round(rowMeans(VP$vals, na.rm = T), 3)*100)
    mycols = rainbow(nrow(VP$vals))
    barplot(VP$vals, ylim = c(0,1), cex.lab = 0.8, main = plot_titles[nrow(VPsummary)], ylab = "Fractions of variance explained", axisnames = FALSE, cex.axis = 0.7, col = mycols, las=2, horiz=FALSE, cex.names = 0.6)
    if(j == 2)
    {legend("topright", xpd=TRUE, legend = c("Time period", "Forest (%)", "Mean annual temp.", "Trap days", "Full sample", "SiteRL", "TimeRL", "SampleRL"), fill = mycols, bty = "n", cex = 0.7)
    }
  }
dev.off()

#Simplified variance partitioning barplot with covariates grouped for Fig. 3

row.names(VPsummary) = c("OTUpres_abs", "OTUabu_cop", "HAPpres_abs", "HAPabu_cop")
write.csv2(VPsummary, "VPsummary.csv")

VPsummary = data.frame(VPsummary)

Climate_incl_Temporal = VPsummary$Time.Period + VPsummary$Random..time + VPsummary$bio1
Spatial = VPsummary$Random..site
SamplingEffort = VPsummary$TrapDays + VPsummary$FullSample
Habitat = VPsummary$ForestWood
SampleID = VPsummary$Random..sample

VPbarplot = t(data.frame(Climate_incl_Temporal, Spatial, SamplingEffort, Habitat, SampleID))
colnames(VPbarplot) = c("OTUpres_abs", "OTUabu_cop", "HAPpres_abs", "HAPabu_cop")
mycols = rainbow(nrow(VPbarplot))
barplot(VPbarplot, ylim = c(0,100), cex.lab = 0.8, ylab = "Fractions of variance explained", axisnames = FALSE, cex.axis = 0.7, col = mycols, las=2, horiz=FALSE, cex.names = 0.6)
legend("topright", xpd=TRUE, legend = rev(rownames(VPbarplot)), fill = rev(mycols), bty = "n", cex = 0.7)
