==> HMSC/5_BetaPlotFigure.R <==
setwd()

library(Hmsc)

#load Hmsc model generated by script 2.
load("models_thin_1000_samples_250_chains_4.Rdata")

modeltype = c("OTU models", "Haplotype models")
plot_titles = c("mOTU occurrence", "mOTU abundance", "Haplotype occurrence", "Haplotype abundance")

#Define functions to calculate what % of taxa had positive and negative responses to each X covariate in the model.
negresponses = function(b) {
  return(length(which(b<0.05)))}
posresponses = function(b) {
  return(length(which(b>0.95)))}

#Beta plots (basis of Fig. 4)
pdf("BetaPlots.pdf")
par(mar = c(6, 4, 3, 2), lwd = 0.1)

pos = list()
neg = list()

for(j in 1:4)
  {
    if(j == 1 | j == 3)
    {plot_main = plot_titles[1:2]}
    if(j == 2 | j == 4)
    {plot_main = plot_titles[3:4]}
  
    m = models[[j]]
    m$covNames = c("Intercept", "Mid-summer", "Late summer", "Off-season", "Forest (%)", "Mean annual temp.", "Trap days", "Full sample")
    postBeta = getPostEstimate(m, parName="Beta")
    pos[[length(pos)+1]] = round(100*apply(postBeta$support, 1, posresponses)/ncol(m$Y),1)
    neg[[length(neg)+1]] = round(100*apply(postBeta$support, 1, negresponses)/ncol(m$Y),1)
    plotBeta(m, post=postBeta, supportLevel = 0.95, param="Sign",
             covNamesNumbers = c(TRUE,FALSE),
             spNamesNumbers=c(FALSE,FALSE),
             cex=c(0.6,0.6,0.8))
    title(main = plot_main[j], cex.main=0.8)
  }
dev.off()


#Numeric summaries of mOTUs with positive responses per covariate:
positive_resps = as.data.frame(pos)
colnames(positive_resps) = c("mOTU occurrence", "mOTU abundance", "Haplotype occurrence", "Haplotype abundance")
rownames(positive_resps) = c("Intercept", "Mid-summer", "Late summer", "Off-season", "Forest (%)", "Mean annual temp.", "Trap days", "Full sample")
positive_resps = positive_resps[-1,]

negative_resps = as.data.frame(neg)
colnames(negative_resps) = c("mOTU occurrence", "mOTU abundance", "Haplotype occurrence", "Haplotype abundance")
rownames(negative_resps) = c("Intercept", "Mid-summer", "Late summer", "Off-season", "Forest (%)", "Mean annual temp.", "Trap days", "Full sample")
negative_resps = negative_resps[-1,]
