==> HMSC/7_ResidualAssociations.R <==
setwd()

library(Hmsc)
library(colorspace)
library(corrplot)

#load Hmsc model generated by script 2.
load("models_thin_1000_samples_250_chains_4.Rdata")

#Residual association plots (basis for Supplementary Figure S10)
#Here plot title combine model name, random effect name and, for spatially or temporally explicit random effects alpha,
#the "typical scale" of autocorrelation and posterior support (p) that alpha>0.

nm = length(models)

pdf("Residual_Association_plots.pdf")
{
  for(j in 1:nm){
      m = models[[j]]
      OmegaCor = computeAssociations(m)
      supportLevel = 0.95
      for (r in 1:m$nr){
        plotOrder = corrMatOrder(OmegaCor[[r]]$mean,order="AOE")
        #plotOrder = 1:m$ns
        toPlot = ((OmegaCor[[r]]$support>supportLevel) + (OmegaCor[[r]]$support<(1-supportLevel))>0)*sign(OmegaCor[[r]]$mean)
        if(m$ns>20){
          colnames(toPlot)=rep("",m$ns)
          rownames(toPlot)=rep("",m$ns)
        }
        mymain = paste0("Associations, ",modelnames[[j]], ": ",names(m$ranLevels)[[r]])
        if(m$ranLevels[[r]]$sDim>0){
          mpost = convertToCodaObject(m)
          alphavals = unlist(poolMcmcChains(mpost$Alpha[[r]][,1]))
          mymain = paste0(mymain,", E[alpha1] = ",round(mean(alphavals),2),", Pr[alpha1>0] = ",round(mean(alphavals>0),2))
        }
        corrplot(toPlot[plotOrder,plotOrder], method = "color",
                 col=colorRampPalette(c("blue","white","red"))(3),
                 mar=c(0,0,1,0),
                 main=mymain,cex.main=0.8)
        me = as.data.frame(OmegaCor[[r]]$mean)
        me = cbind(m$spNames,me)
        colnames(me)[1] = ""
        po = as.data.frame(OmegaCor[[r]]$support)
        po = cbind(m$spNames,po)
        colnames(po)[1] = ""
        ne = as.data.frame(1-OmegaCor[[r]]$support)
        ne = cbind(m$spNames,ne)
        colnames(ne)[1] = ""
        vals = list("Posterior mean"=me,"Pr(x>0)"=po,"Pr(x<0)"=ne)
        }}}
dev.off()
